refactoring structure
