+++ App.tsx
@@ -8,7 +8,7 @@ import { useFonts as useLato, Lato_400Regular } from "@expo-google-fonts/lato";
 
 import { theme } from "./src/theme/theme";
 
-import Restaurants from "./src/screens/Restaurants";
+import Restaurants from "./src/screens/Restaurants/Restaurants";
 
 export default function App(): JSX.Element {
   let [oswaldLoaded] = useOswald({
