+++ package.json
@@ -12,15 +12,18 @@
   "dependencies": {
     "@expo-google-fonts/lato": "^0.2.3",
     "@expo-google-fonts/oswald": "^0.2.3",
+    "@react-navigation/bottom-tabs": "^7.2.0",
+    "@react-navigation/native": "^7.0.14",
     "babel-plugin-inline-import": "^3.0.0",
     "eslint-plugin-prettier": "5.0.0",
-    "expo": "^51.0.14",
+    "expo": "^51.0.38",
     "expo-font": "~12.0.7",
     "expo-status-bar": "~1.12.1",
     "react": "18.2.0",
     "react-native": "0.74.2",
     "react-native-paper": "^5.11.6",
-    "react-native-safe-area-context": "4.10.1",
+    "react-native-safe-area-context": "4.10.5",
+    "react-native-screens": "3.31.1",
     "react-native-svg": "15.2.0"
   },
   "devDependencies": {
@@ -31,5 +34,6 @@
     "prettier": "^3.1.1",
     "typescript": "^5.1.3"
   },
-  "private": true
+  "private": true,
+  "packageManager": "yarn@1.22.22+sha512.a6b2f7906b721bba3d67d4aff083df04dad64c399707841b7acf00f6b133b7ac24255f2652fa22ae3534329dc6180534e98d17432037ff6fd140556e2bb3137e"
 }
