+++ src/components/PlaceCard/PlaceCard.tsx
@@ -1,18 +1,8 @@
-import { Image, StyleSheet, View } from "react-native";
+import { Image, View } from "react-native";
 import React from "react";
 import { Card, Text } from "react-native-paper";
-import { theme } from "../theme/theme";
 import { SvgXml } from "react-native-svg";
-
-type PlaceCardProps = {
-  name: string;
-  icon: string;
-  photos: string[];
-  address: string;
-  isOpenNow: boolean;
-  rating: number;
-  isClosedTemporarily: boolean;
-};
+import { styles, PlaceCardProps } from "./placeCardProps";
 
 const PlaceCard = ({
   name,
@@ -57,48 +47,6 @@ const PlaceCard = ({
 };
 
 export default PlaceCard;
-
-const styles = StyleSheet.create({
-  container: {
-    paddingHorizontal: "5%",
-  },
-  card: {
-    padding: theme.spacing.md,
-  },
-  cover: {
-    width: "90%",
-    alignSelf: "center",
-  },
-  title: {
-    paddingTop: theme.spacing.lg,
-    fontFamily: theme.fonts.fontFamily.heading,
-    fontSize: theme.fonts.fontSize.h5,
-  },
-  address: {
-    fontFamily: theme.fonts.fontFamily.body,
-    fontSize: theme.fonts.fontSize.caption,
-  },
-  ratingContainer: {
-    flexDirection: "row",
-    paddingVertical: theme.spacing.sm,
-  },
-  openContainer: {
-    flexDirection: "row",
-    flexGrow: 1,
-    justifyContent: "flex-end",
-    paddingRight: theme.spacing.md,
-  },
-  closedText: {
-    color: "red",
-    marginHorizontal: theme.spacing.md,
-    fontFamily: theme.fonts.fontFamily.monospace,
-  },
-  icon: {
-    width: 15,
-    height: 15,
-    marginLeft: theme.spacing.md,
-  },
-});
 
 const starPath = `<?xml version="1.0" encoding="iso-8859-1"?>
 <!-- Generator: Adobe Illustrator 19.0.0, SVG Export Plug-In . SVG Version: 6.00 Build 0)  -->
