+++ src/components/SearchBar/searchbarProps.ts
@@ -1,0 +1,7 @@
+import { StyleSheet } from "react-native";
+
+export const styles = StyleSheet.create({
+  searchbar: {
+    marginHorizontal: "5%",
+  },
+});
