+++ src/constants/routeConstants.ts
@@ -1,0 +1,3 @@
+export const RouteConstants = {
+  homeScreen: "homeScreen",
+};
