+++ src/screens/Restaurants/Restaurants.tsx
@@ -1,16 +1,10 @@
 import { StatusBar as ExpoStatusBar } from "expo-status-bar";
-import {
-  StyleSheet,
-  View,
-  SafeAreaView,
-  StatusBar,
-  FlatList,
-} from "react-native";
+import { View, SafeAreaView, FlatList } from "react-native";
 import React from "react";
 
-import Searchbar from "../components/Searchbar";
-import PlaceCard from "../components/PlaceCard";
-import { theme } from "../theme/theme";
+import Searchbar from "../../components/SearchBar/Searchbar";
+import PlaceCard from "../../components/PlaceCard/PlaceCard";
+import { styles } from "./restrauntsProps";
 
 const data = [
   {
@@ -53,7 +47,7 @@ const data = [
 
 // type RestaurantsProps = {};
 
-const Restaurants = (): JSX.Element => {
+function Restaurants(): JSX.Element {
   return (
     <>
       <SafeAreaView style={styles.container}>
@@ -82,26 +76,6 @@ const Restaurants = (): JSX.Element => {
       <ExpoStatusBar style="auto" />
     </>
   );
-};
+}
 
 export default Restaurants;
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    marginTop: StatusBar.currentHeight,
-  },
-  search: {
-    flexGrow: 0.05,
-    justifyContent: "center",
-    paddingLeft: theme.spacing.lg,
-  },
-  list: {
-    flexGrow: 0.95,
-    backgroundColor: theme.colors.bg.primary,
-    paddingTop: theme.spacing.md,
-  },
-  listItemContainer: {
-    paddingVertical: theme.spacing.md,
-  },
-});
